move QR code into reusable view
